Updated Player Class and Removed Console Logs in Script.js
The script.js file has been updated to include the constructor, draw, and update methods in the Player class. This allows the player's position and size to be defined and updated. The console.log statements have been removed from the InputHandler class. The animate function has also been updated to call the draw and update methods of the player class.