Key Event and Movement Update in JavaScript
The changes include swapping the 'keydown' event listener for 'keyup' in the script.js file. Additionally, a console log was added to track key events. The player's horizontal movement was also updated to respond to the "ArrowRight" key. The typo in the class name 'Backgroud' was corrected to 'Background'. Lastly, the 'update' method in the 'animate' function now includes 'input' as a parameter.